chore(js/client): moved client to beta namespace
