+++ js/genkit/package.json
@@ -151,7 +151,7 @@
       "import": "./lib/plugin.mjs",
       "default": "./lib/plugin.js"
     },
-    "./client": {
+    "./beta/client": {
       "types": "./lib/client/index.d.ts",
       "require": "./lib/client/index.js",
       "import": "./lib/client/index.mjs",
@@ -214,7 +214,7 @@
       "plugin": [
         "lib/plugin"
       ],
-      "client": [
+      "beta/client": [
         "lib/client/index"
       ]
     }
